refactor(shared): rename sandbox api key helpers
