+++ packages/shared/lib/utils/encryption.manager.integration.test.ts
@@ -5,7 +5,7 @@ import db, { multipleMigrations } from '@nangohq/database';
 import encryptionManager, { EncryptionManager } from './encryption.manager.js';
 import { seedAccountEnvAndUser } from '../seeders/index.js';
 import environmentService from '../services/environment.service.js';
-import { decryptSandboxSigningSecret } from '../services/sandbox-api-key.service.js';
+import { decryptSandboxSigningSecret } from '../services/sandbox-api-key.js';
 import secretService from '../services/secret.service.js';
 
 import type { DBCustomerKey } from '@nangohq/types';
